+++ src/fileio/fileio.h
@@ -12,7 +12,7 @@ using namespace std;
 namespace FileIO
 {
     void store_char_vector(vector<char> &vec, string path);
-    bool read_char_vector(vector<char> &vec, string path, int required_size);
+    bool read_char_vector(vector<char> &vec, string path, long long required_size);
     void store_unordered_map(unordered_map<long long, char> &map, string path);
     bool read_unordered_map(unordered_map<long long, char> &map, string path, int required_size);
 };
